Курсовая. Высокая сложность

+++ src/edu/skypro/coursework/Employee.java
@@ -5,11 +5,11 @@ import java.util.Objects;
 public class Employee {
     private final String fullName;
     private int department;
-    private int salary;
+    private double salary;
     private final int id;
     static int count;
 
-    public Employee(String fullName, int department, int salary) {
+    public Employee(String fullName, int department, double salary) {
         this.fullName = fullName;
         this.department = department;
         this.salary = salary;
@@ -24,7 +24,7 @@ public class Employee {
         return department;
     }
 
-    public int getSalary() {
+    public double getSalary() {
         return salary;
     }
 
@@ -36,7 +36,7 @@ public class Employee {
         this.department = department;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
